requirements.txt
flask
librosa
numpy
gunicorn
openai-whisper
git+https://github.com/openai/whisper.git
soundfile
io 